rename internal/config -> internal/rsyncdconfig
